Finalizando o Projeto

+++ src/cache/cache-del/cache-del.service.ts
@@ -4,8 +4,11 @@ import { RedisClient } from '../redis/RedisClient';
 
 @Injectable()
 export class CacheDelService extends RedisClient{
-    async del(keys ?:Array<RedisKey>) :Promise<number>{
-       const deletions =  await this.client.del(keys)
-       return deletions
+    async del(key :RedisKey,keys ?:Array<RedisKey>) :Promise<void>{
+        if(keys){
+        await this.client.del(keys)
+        }else {
+        await this.client.del(key)
+        }
     }
 }
